api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich[toml]>=0.9.2",
    "bcrypt==4.0.1",
    "boto3>=1.42.14",
    "boto3-stubs~=1.42.14",
    "cryptography>=46.0.3",
    "fastapi>=0.121.3",
    "fastmcp>=2.14.3",
    "openai>=2.14.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "tortoise-orm[aiomysql]>=0.25.3",
    "uvicorn>=0.40.0",
]
[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.aerich]
tortoise_orm = "databases.db_config"
location = "./migrations"
src_folder = "./."
